+++ dids/src/main/kotlin/web5/sdk/dids/Serialization.kt
@@ -1,6 +1,5 @@
 package web5.sdk.dids
 
-import com.fasterxml.jackson.annotation.JsonInclude
 import com.fasterxml.jackson.core.JsonGenerator
 import com.fasterxml.jackson.core.JsonParser
 import com.fasterxml.jackson.core.JsonProcessingException
@@ -8,53 +7,11 @@ import com.fasterxml.jackson.databind.DeserializationContext
 import com.fasterxml.jackson.databind.JsonDeserializer
 import com.fasterxml.jackson.databind.JsonNode
 import com.fasterxml.jackson.databind.JsonSerializer
-import com.fasterxml.jackson.databind.ObjectMapper
-import com.fasterxml.jackson.databind.ObjectWriter
-import com.fasterxml.jackson.databind.SerializationFeature
 import com.fasterxml.jackson.databind.SerializerProvider
-import com.fasterxml.jackson.module.kotlin.registerKotlinModule
 import com.nimbusds.jose.jwk.JWK
 import web5.sdk.dids.didcore.Purpose
 import java.io.IOException
 
-/**
- * A singleton for json serialization/deserialization, shared across the SDK as ObjectMapper instantiation
- * is an expensive operation.
- * - Serialize ([stringify])
- *
- * ### Example Usage:
- * ```kotlin
- * val offering = Json.objectMapper.readValue<Offering>(payload)
- *
- * val jsonString = Json.stringify(myObject)
- *
- * val node = Json.parse(payload)
- * ```
- */
-public object Json {
-  /**
-   * The Jackson object mapper instance, shared across the lib.
-   *
-   * It must be public in order for typed parsing to work as we cannot use reified types for Java interop.
-   */
-  public val jsonMapper: ObjectMapper = ObjectMapper()
-    .registerKotlinModule()
-    .findAndRegisterModules()
-    .setSerializationInclusion(JsonInclude.Include.NON_NULL)
-    .disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS)
-
-  private val objectWriter: ObjectWriter = jsonMapper.writer()
-
-  /**
-   * Converts a kotlin object to a json string.
-   *
-   * @param obj The object to stringify.
-   * @return json string.
-   */
-  public fun stringify(obj: Any): String {
-    return objectWriter.writeValueAsString(obj)
-  }
-}
 
 /**
  * Serialize JWK into String.
@@ -81,7 +38,7 @@ public class JwkDeserializer : JsonDeserializer<JWK>() {
 }
 
 /**
- * Deserialize String into JWK.
+ * Deserialize String into List of Purpose enums.
  *
  */
 public class PurposesDeserializer : JsonDeserializer<List<Purpose>>() {
